Correct way to push to github
